requirements.txt
django
djangorestframework
psycopg2-binary
celery
redis
pandas
openpyxl
python-dateutil
